create util for navigator functions and refactor AR to use it
