+++ app/components/Map.js
@@ -16,6 +16,7 @@ import targetImg from '../assets/blackPin.png';
 import { PinCallout } from './PinCallout';
 import PinEditButton from './PinEditButton';
 import { myCurrLoc, currLoc } from '../lib/db/db';
+import * as geoAction from './utils';
 
 
 export default class Map extends Component {
@@ -28,8 +29,8 @@ export default class Map extends Component {
       loaded: false,
       friendLocs: {},
       stateLocation: {
-        longitude: null,
-        latitude: null,
+        longitude: 0,
+        latitude: 0,
         longitudeDelta: 0.005,
         latitudeDelta: 0.005
       }
@@ -150,7 +151,7 @@ export default class Map extends Component {
   renderFriends() {
     const { friends } = this.props;
     let copy = this.state.friendLocs;
-    
+
     // renders friends current locations
     return _.map(copy, (coords, id) => {
         return (
@@ -199,8 +200,8 @@ export default class Map extends Component {
         <MapView
           ref="map"
           showsUserLocation={true}
-          initialRegion={{ longitudeDelta: 0.005, latitude: currLoc.latitude,longitude: currLoc.longitude, latitudeDelta: 0.005 }}
-          region={this.state.position}
+          initialRegion={stateLocation}
+          region={stateLocation}
           style={styles.map}
           showsCompass={true}
           onLongPress={ (e) => {
@@ -209,7 +210,7 @@ export default class Map extends Component {
             }
           }
         >
-        
+
         { Object.keys(pins).length !== 0 ? this.renderMarkers.call(this) : void 0 }
 
         { this.state.loaded === true ? this.renderFriends.call(this) : void 0 }
